rename polynomial.h/cpp
