+++ source/core/calibration/CMakeLists.txt
@@ -4,7 +4,7 @@ set(SOURCES
   ${dir}/parameter.cpp
   ${dir}/coef_function.cpp
   ${dir}/coef_function_factory.cpp
-  ${dir}/polynomial.cpp
+  ${dir}/Polynomial.cpp
   ${dir}/calibration.cpp
   )
 
@@ -12,7 +12,7 @@ set(HEADERS
   ${dir}/parameter.h
   ${dir}/coef_function.h
   ${dir}/coef_function_factory.h
-  ${dir}/polynomial.h
+  ${dir}/Polynomial.h
   ${dir}/calibration.h
   )
