+++ tests/core/calibration/CMakeLists.txt
@@ -4,7 +4,7 @@ set(SOURCES
   ${dir}/parameter.cpp
   ${dir}/coef_function.cpp
   ${dir}/coef_function_factory.cpp
-  ${dir}/polynomial.cpp
+  ${dir}/PolynomialTest.cpp
   ${dir}/calibration.cpp
   )
