+++ tests/core/calibration/calibration.cpp
@@ -1,7 +1,7 @@
 #include "gtest_color_print.h"
 #include <core/calibration/calibration.h>
 #include <core/calibration/coef_function_factory.h>
-#include <core/calibration/polynomial.h>
+#include <core/calibration/Polynomial.h>
 
 class CalibID : public TestBase
 {
